% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oplspvs.R
\name{oplspvs}
\alias{oplspvs}
\title{R orthogonal projections of latent structures with permutation over variable selection}
\usage{
oplspvs(
  directory_of_roplspvs,
  directory_of_analysis,
  projectname,
  date_of_analysis,
  filename_matrix,
  decimal_separator,
  filename_sampleID,
  colname_groupID,
  groupsnumeric,
  colname_secID,
  no_permutations_sans_vs,
  no_permutations_post_vs,
  no_permutations_post_vs_selected_models,
  no_permutations_over_vs,
  p_pearson_of_pcorr_cutoff,
  setseedfirstmodel,
  order_of_groups,
  comparisons_to_run,
  foldername_Rmarkdownfiles,
  foldername_of_input_matrix_and_sampleID,
  foldername_output_reports,
  directory_input_matrix_sampleID,
  directory_output_reports,
  filename_Rmarkdownfile_each_model,
  filename_Rmarkdownfile_summary,
  filename_function_file,
  replace_0,
  filter_percent_in_each_group,
  replace_NA,
  log_transform,
  pcorr_cutoff_Model1_joint_models,
  pcorr_cutoff_Model1_stratified_models,
  variable_selection_using_VIP,
  no_of_ortho_pre_vs_Model1_joint_models,
  no_of_ortho_pre_vs_Model1_stratified_models,
  no_of_ortho_post_vs_Model1_joint_models,
  no_of_ortho_post_vs_Model1_stratified_models,
  max_no_of_ortho_pre_vs,
  max_no_of_ortho_post_vs,
  prefered_pR2_and_pQ2_permutated_post_vs,
  pcorr_diff,
  variable_names_length,
  variable_names_position,
  cluster,
  name_intermediate_dir,
  each_model_or_summary,
  model_strategies_to_run
)
}
\arguments{
\item{directory_of_roplspvs}{path to repository cloned from https://github.com/MarikaStrom/roplspvs}

\item{directory_of_analysis}{path to folder where the outputR folder with results are created}

\item{projectname}{"projectname" will appear in filenames and header of reports with underscores removed.}

\item{date_of_analysis}{"yymmdd" Date of analysis will appear in filenames.}

\item{filename_matrix}{"filename.txt" file containing your data to be modeles. subjectID in the first column as rownames and variable names in the first row as columnnames. All names must be unique.}

\item{decimal_separator}{"dot" or "comma" as decimal separator in filename_matrix}

\item{filename_sampleID}{"filename.txt" file containing metadata with sampleID in the first column and containing one column with the groups to be compared and one column with secondaryID (for example gender) if stratification is desired. All subjectIDs has to be unique and ordered in the same way as in datamatrix.}

\item{colname_groupID}{"column name"* of groups to compare in sampleID file}

\item{groupsnumeric}{"yes" if you want tables in summary file to be sorted by numeric group belonging. "no" if not.}

\item{colname_secID}{"column name"* column name of secondary id in sampleID file to stratify or write "joint" for no stratification}

\item{no_permutations_sans_vs}{Numeric. Number of permutations sans variable selection.}

\item{no_permutations_post_vs}{Numeric. Number of permutations post variable selection during model selection.}

\item{no_permutations_post_vs_selected_models}{Numeric. Number of permutations post variable selection in selected models.}

\item{no_permutations_over_vs}{Numeric. Number of permutations including variable selection in selected models.}

\item{p_pearson_of_pcorr_cutoff}{Numeric. P-value for p(corr) cutoff during variable selection in model 2 and minimum cutoff in model 3 and 4. Optionally in model 1.}

\item{setseedfirstmodel}{Numeric. Setseed of the first model. Second model will have setseedfirstmodel+1 etc.}

\item{order_of_groups}{Character vector or numeric vector containing correct order of groups or enter "correct" if order of levels in colname_groupID is already correct. Deseased first and controls last. This will define dirction of scores as high in diseased.}

\item{comparisons_to_run}{Numeric vector indicating models to run if all models are to be run enter "all"}

\item{foldername_Rmarkdownfiles}{"foldername" Folder where .R and .rmd files required by the script are stored. It is located in the roplspvs folder.}

\item{foldername_of_input_matrix_and_sampleID}{"foldername" Folder where datamatrix and sampleID files are stored.}

\item{foldername_output_reports}{"foldername"  Folder where one .html file and one .Rdata for each group comparison and a summary .html and a .Rdata file of all comparisons will be created. Folder is automatically created if it does not excist.}

\item{directory_input_matrix_sampleID}{"path" Path to where "filename_matrix" and "filename_sampleID" files are stored.}

\item{directory_output_reports}{"path" Path to where "output_reports" are created in foldername_output_reports.}

\item{filename_Rmarkdownfile_each_model}{"filename.Rmd" rendered by function once per comparison to produce html report and .Rdata file for each comparison}

\item{filename_Rmarkdownfile_summary}{"filename.Rmd" rendered by function if "each_model_or_summary" is set to "both" or "summary" after all comparisons have been performed to produce a summary tables of all models.}

\item{filename_function_file}{"filename.R" All functions required are saved in this file.}

\item{replace_0}{F or "lld" or a value to replace 0 values with. Replacing is performed before filtering. Lld (lower limit of detection) is calculated by llq/3 with llq (lower limit of quantification) is the lowest value in the dataset.}

\item{filter_percent_in_each_group}{Numeric. Missing value tolerance in each group which are compared.}

\item{replace_NA}{F or "lld" or a value to replace NA with. If not replaced NAs will be imputed using NIPAL by ropls package. Replaceing is performed after filtering. Lld (lower limit of detection) is calculated by llq/3 with llq (lower limit of quantification) is the lowest value in the dataset.}

\item{log_transform}{T or F if datamatrix is transformed by natural logarithm}

\item{pcorr_cutoff_Model1_joint_models}{P(corr) cutoff in model1 for joint models. Either enter a value with p(corr)cutoff for all joint comparisons or vector containing p(corr) cutoff for each comparison in model_table_to_analyse or enter "according to p-value" to generate p(corr) corresponding to selected pvalue entered in p_pearson_of_pcorr_cutoff.}

\item{pcorr_cutoff_Model1_stratified_models}{P(corr) cutoff in model1 for stratified models. Either enter a value with p(corr)cutoff for all stratified comparisons or vector containing p(corr) cutoff for each comparison in model_table_to_analyse or enter "according to p-value" to generate p(corr) corresponding to selected pvalue entered in p_pearson_of_pcorr_cutoff}

\item{variable_selection_using_VIP}{"yes" if VIP is used during variable selection and "no" if only p(corr) is used.}

\item{no_of_ortho_pre_vs_Model1_joint_models}{Numeric. Number of orthogonal variables in model pre variable selection of Model strategy 1 for joint models. Either enter a value for nonstratified comparisons or a vector containing the number of orthogonal variables investigated in model pre variable selection for each model in model_table_to_analyse.}

\item{no_of_ortho_pre_vs_Model1_stratified_models}{Numeric. Number of orthogonal variables in model pre variable selection of Model strategy 1 for stratified models. Either enter a value for stratified comparisons or vector containing the amount of orthogonal variables investigated in model pre variable selection for each model in model_table_to_analyse.}

\item{no_of_ortho_post_vs_Model1_joint_models}{Numeric. Number of orthogonal variables in model post variable selection of Model strategy 1 for joint models. Either enter a value for joint comparisons or vector containing the amount of orthogonal variables investigated in model post variable selection for each model in model_table_to_analyse.}

\item{no_of_ortho_post_vs_Model1_stratified_models}{Numeric. Number of orthogonal variables in model post variable selection of Model strategy 1 for stratified models. Either enter a value for stratified comparisons or vector containing the amount of orthogonal variables investigated in model post variable selection for each model in model_table_to_analyse.}

\item{max_no_of_ortho_pre_vs}{Numeric. Max number of orthogonal variables in model pre variable selection}

\item{max_no_of_ortho_post_vs}{Numeric. Max number of orthogonal variables in model post variable selection}

\item{prefered_pR2_and_pQ2_permutated_post_vs}{Numeric. Prefered pR2 and pQ2 determined by permutation post variable selection during model optimization. The lower value the larger weight will be given to pQ2 instead of Q2 as well as instead of diff between R2 and Q2. THe value is applied during pcorr optimization in model strategy 3 and 4 and during selection of othogonals in model strategy 2 to 4 and while selecting best iteration model in model strategy 5. Minimum value is 1/no_permutations_post_vs.}

\item{pcorr_diff}{Numeric. A value by which pcorr cutoff is decreased to allow more variables in the model if Q2 is increased more than 1 percent. Used in model strategy 4.}

\item{variable_names_length}{Numeric. Number of characters of variablenames shown in loadingplot or "all"}

\item{variable_names_position}{"beginning", "end" or "all" give part of string of variablenames shown in loadingplot}

\item{cluster}{"no" if analysis is run locally and "yes" if script is run using the bash scripts to run the R-scripts passing modelnumbers to run as arguments ("args").}

\item{name_intermediate_dir}{"name_of_directory" enter name of intermediate directory. On UPPMAX it is "SNIC_TMP" else enter "default"}

\item{each_model_or_summary}{"summary" if only summary should be run and "each" if only models of each comparison should be run. "both" if first running each comparison followed by summary. "both" is not possible if cluster is "yes". "summary" requires that .Rdata files from "each" exists in outputR folder.}

\item{model_strategies_to_run}{Numeric vector indicating which model strategies to run. Model strategy 3 requires that 2 is run. Model strategy 4 requires that 3 and 2 is run.}
}
\value{
Outputs in a folder called outputR in your analysis folder. one .html file and one .Rdata file
per group comparison  containing five models and one summary html file containing
tables of all models of all comparisons. Also .txt files containing tables with loadings are created.
}
\description{
Pipeline for OPLS models, variable selection and permutations pre, post and over variable selection
}
\details{
Performs opls models comparing all groups to eachother, performs variable selection using 5 strategies
all based on p(corr) and optionally VIP. Produces scores plots, loading plots, rock curves and significant
testing. In addition to permutation pre and post variable
selection included in the ropls package also permutations over variable selection with proceeding
variable selection of every permutation resulting in p-values for R2 and Q2 including the variable
selection procedure. It produces tables of all group comparisons including optional stratification
by metadata.

In filename_sampleID and in filename_matrix do not use the following symbols in sampleID or variable names;
?, %, ^, &, *, (, ), -, #, ?, ,, <, >, /, |, , ], {, }, [, and $
Missing values should be indicated by "", "NA" or "Inf"
}
\examples{

## The function may be run using the workflow setting the parameters in the files
## roplspvs_Configure_Get_Started.R and roplspvs_Configure_Advanced.R. These files are
## saved in your project folder where also a folder namned data containing your datamatrix
## and metadata matrix is saved. This is where your result folder outputR is generated. To run example data:

setwd("<path to your project folder>")
source("<path to roplspvs>/Roplspvs_Run.R")

## The workflow runs the following function:
oplspvs(directory_of_roplspvs,directory_of_analysis,
projectname, date_of_analysis, filename_matrix, decimal_separator, filename_sampleID, colname_groupID, groupsnumeric,
colname_secID, no_permutations_sans_vs, no_permutations_post_vs, no_permutations_post_vs_selected_models, no_permutations_over_vs,
p_pearson_of_pcorr_cutoff, setseedfirstmodel, order_of_groups, comparisons_to_run, model_strategies_to_run,
foldername_Rmarkdownfiles, foldername_of_input_matrix_and_sampleID, foldername_output_reports,
directory_input_matrix_sampleID, directory_output_reports, filename_Rmarkdownfile_each_model, filename_Rmarkdownfile_summary,
filename_function_file, replace_0, filter_percent_in_each_group, replace_NA, log_transform, pcorr_cutoff_Model1_joint_models, pcorr_cutoff_Model1_stratified_models,
variable_selection_using_VIP, no_of_ortho_pre_vs_Model1_joint_models, no_of_ortho_pre_vs_Model1_stratified_models,
no_of_ortho_post_vs_Model1_joint_models, no_of_ortho_post_vs_Model1_stratified_models, max_no_of_ortho_pre_vs, max_no_of_ortho_post_vs,
prefered_pR2_and_pQ2_permutated_post_vs, pcorr_diff, variable_names_length, variable_names_position,
cluster, name_intermediate_dir, each_model_or_summary, model_strategies_to_run)
}
